Issue #4080: Split and Organize Checkstyle inputs by Test for SingleSpaceSeparator
