+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/singlespaceseparator/InputSingleSpaceSeparatorComments.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.whitespace;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.singlespaceseparator;
 
-public class InputSingleSpaceComments {
+public class InputSingleSpaceSeparatorComments {
     /* always correct */ int i = 0;
     int   /* wrong if X is enabled */     j = 0;
     int k;   // Multiple whitespaces before comment
